requirements.txt
streamlit
tensorflow
numpy
pillow
reportlab
datetime
